COM-1197-sidemenu corrections following feedbacks
