+++ src/modules/client/layouts/Layout.vue
@@ -23,8 +23,8 @@
           <q-separator :key="`separator-${route.ref}`" />
           </template>
         </template>
-        <ni-side-menu-footer :label="headerFormat.label" :userId="loggedUser._id" :interface-type="interfaceType"
-          @myClick="headerFormat.myClick"/>
+        <ni-side-menu-footer :label="footerLabel" :userId="loggedUser._id" :interface-type="CLIENT"
+          @click="connectToBotMessenger"/>
       </q-list>
       <div :class="chevronContainerClasses">
         <q-btn :class="chevronClasses" dense round unelevated :icon="menuIcon" @click="isMini = !isMini" />
@@ -40,21 +40,34 @@
 </template>
 
 <script>
+import { Cookies } from 'quasar';
 import { layoutMixin } from '@mixins/layoutMixin';
 import { sideMenuMixin } from '@mixins/sideMenuMixin';
 import SideMenuFooter from '@components/menu/SideMenuFooter';
 import MenuItem from '@components/menu/MenuItem';
-import { menuItems } from './menuItems'
+import { CLIENT } from '@data/constants';
+import { menuItemsMixin } from '../mixins/menuItemsMixin'
 
 export default {
   name: 'ClientLayout',
-  mixins: [layoutMixin, sideMenuMixin, menuItems],
+  mixins: [layoutMixin, sideMenuMixin, menuItemsMixin],
   components: {
     'ni-side-menu-footer': SideMenuFooter,
     'ni-menu-item': MenuItem,
   },
+  data () {
+    return {
+      CLIENT,
+    }
+  },
   mounted () {
     this.collapsibleOpening();
+  },
+  methods: {
+    connectToBotMessenger () {
+      const token = Cookies.get('alenvi_token');
+      window.location.href = `${process.env.MESSENGER_LINK}?ref=${token}`
+    },
   },
 }
 </script>
